Ifpack2: AdditiveSchwarz: Fixed inner preconditioner creation.
The code in setup() that ensures creation of the preconditioner (if
necessary) wasn't giving the inner matrix to the inner preconditioner.
Instead, it was passing Teuchos::null to the inner preconditioner, then
calling the inner preconditioner's setMatrix() method (by calling
setInnerPreconditioner()).  This is unnecessary and requires an
additional initialize() and compute() call on the inner preconditioner.
This commit avoids the unnecessary redundant calls and also ensures that
the inner preconditioner gets the new inner matrix, even if the inner
preconditioner already exists.

I also added a private method to AdditiveSchwarz so that it can get the
current name of the inner preconditioner.  This indirection will
facilitate the eventual addition of the inner preconditioner's name to
AdditiveSchwarz's ParameterList.

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => passed: passed=23,notpassed=0 (0.52 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (0.08 min)
Other local commits for this build/test group: ab87370