TPetra/KokkosRefactor: added nonmeber createCopy function; describe output formatted
Since the copy constructor of MV is using View Semantics an explicit createCopy function
was needed. Added that for both MultiVector and Vector.
Mark changed some formatting for the describe function.
Made sure that the Kokkos::DualView is initialized with the correct number of columns, even
if the number of rows is zero. This fixes a MatVec test with a Matrix with zero NNZ.