TPetra: work on View semantics
Changed operator= of the refactored MultiVector to be largely View Semantics
Attention: whichVectors_ as well as some members of the base class DistObject
are still copied though map_ is view semantics since it is an RCP.

Added createCopy non member functions to the original vector and multivector
class. So in order to have the original copy constructor behaviour one can now
do:
   vector a = ....
   vector b(a);
   b = createCopy(a);
This will work with the original class as well as the refactored classes.
For the original classes to deep_copies will occure (in the copy constructor,
and again for operator=.

Changed an Import unit_test to use this mechanism.