Ifpack2: Diagonal: Removed unimplemented applyMat method.
For some reason, Diagonal had a method, applyMat, which was
unimplemented.  Its documentation claims that it applies the original
input matrix A to a multivector.  This method was not inherited from any
base class of Diagonal.  It's also unnecessary, since one may call
getMatrix() on the Diagonal instance and apply it oneself.  Thus, I'm
removing it.