Ifpack2: Refactored and improved documentation of internals of RILUK.
Refactored interface of IlukGraph, so that it looks more like the other
Ifpack2 preconditioners (even though it's not a preconditioner itself).
Added Doxygen documentation to RILUK and IlukGraph header files (so that
Doxygen will generate their class documentation).  Improved class
documentation of RILUK and IlukGraph.  Reformatted code to improve
readability.