MueLu: Hierarchy: fixing a dependency problem in the Setup
Was trying to switch the Simple Emin test back on, but encountered this
problem.

Recent (few weeks ago) changes to handle smoother dependencies properly
exposed this problem. Some coarse solvers may require some other data
besides "A". For instance, ProjectorSmoother requires both "A" and
"Nullspace". However, "Nullspace" was erased after the coarseRAP.Build()
call as we did the coarse solver DeclareInput too late to preserve the
data. The typical trace in this situation looked like this:

 MueLu::Level(0)::GetFactory(Aggregates, 0): No FactoryManager
   during request for data "     Aggregates" on level 0 by factory TentativePFactory
   during request for data "              P" on level 1 by factory EminPFactory
   during request for data "              P" on level 1 by factory TransPFactory
   during request for data "              R" on level 1 by factory RAPFactory
   during request for data "              A" on level 1 by factory TentativePFactory
   during request for data "      Nullspace" on level 2 by factory NullspaceFactory
   during request for data "      Nullspace" on level 2 by factory NullspacePresmoothFactory
   during request for data "      Nullspace" on level 2 by factory ProjectorSmoother
   during request for data "    PreSmoother" on level 2 by factory NoFactory

I tried to move the requests of smoothers and direct solvers before
calling the Build method. It seems to work, but there are several
potential issues there:

  1. We may need to construct more data than is truly necessary, due to
     trying to satisfy both smoother and direct solver dependencies.

  2. It still does not help if the data we want is too exotic for anyone
     else to want it.

I don't think 1) poses a problem at our selection of algorithms. 2) is
much more serious issue, and is much more general. It seems that all the
Hierarchy data falls loosely into two categories: the data which we
*need* on all levels, and the data which we may *want* on some level.
The first category contains standard culprits: A, R, P, Aggregates, ...
The second category contains Coordinates, or Nullspace on the coarsest
level. The current implementation of the Hierarchy is poorly suited for
the second category, as we deal with at most three levels. That means
that if we suddenly need coordinates on level 5, there is no way to
construct them. Currently, we hack around this by using RAP transfer
factories to transfer optional stuff.

The optional data problem comes up once in a while. One of the possible
solutions would be to keep all factory managers and all the data until
the end of the construction, and then do the clean up. That would allow
us to transfer, say, coordinates from level 0 to level 5 if we want to.
I don't know how expensive this idea is, or why the original
implementation was chosen to deal with only three levels.