Tpetra: remove node equality tests from Matrix reader
The old code was trying to compare the node objects in row map with that
in domain and range maps by comparing the object pointers. We don't
actually need to compare the objects, we simply need to have the same
node type. But that is already done in the function declaration, as it
is templated on the map_type which contains node type. So it would not
even compile with different node types.