Tpetra:  CMake logic to automatically generate .cu files.
With the new Kokkos, it is impossible to separate CUDA from regular C++
code, and thus everything needs to be compiled with NVCC if
KokkosRefactor and CUDA are enabled.  To support ETI in this scenario,
I've added CMake logic to generate .cu files for each .cpp file at
configure time that includes the .cpp allowing it to be compiled by
NVCC.  I take the corresponding .cpp file out of the SOURCES list so it
is no longer compiled by the regular compiler.  This simplifies things
so that no preprocessor logic is needed.

Unfortunately ETI still doesn't quite work due to several errors in the
new Kokkos kernels that are only exposed when everything is instantiated
through ETI.