Fix issue with excluding too much from the package source
A change to the cpack system to support external repos changed some of the
regular expressions that are used to exclude files and directories from the base
level of Trilinos. When Trilnos is the primary project the base dir was set to
"" which caused regular expressions that had been "<base dir>/doc" to be "/doc"
which matched a lot more than was intended. This fixes that so that when
Trilinos is the primary project the base dir is set to the correct path.