MueLu: fixing memory problem
This commit fixes a problem introduced by commit 1df04658 (added a new
"MaxNodesPerAggregate" parameter).  The issue is that the MaxLinkPhase
was computing aggregate sizes and then altering them.  This then showed
up as invalid memory reads in the CoordinatesTransferFactory, and
resulted in botched coarse grid coordinates.

The fix is to change the default behavior of ComputeAggregateSizes to
always recompute the aggregate sizes.  Thanks to Andrey for reviewing a
preliminary patch and suggesting improvements.

ComputeAggregateSizes now takes two boolean arguments.  The first
(default=true) determines whether to force recomputation.  The second
(default=false) determines whether the computed sizes are then cached.

I traced the issue using this test:

mpirun -np 15 valgrind ./MueLu_ScalingTestParamList.exe --nx=36 --ny=36
--nz=36 --matrixType=Laplace3D  --linAlgebra=Tpetra --xml=saZ1rcb.xml
--notimings --solver=none

and these settings in the input deck for the repartitioning.

    <ParameterList name="myRepartitionFact">
      <Parameter name="factory" type="string" value="RepartitionFactory"/>
      <Parameter name="A" type="string" value="myRAPFact"/>
      <Parameter name="Partition" type="string" value="myZoltanInterface"/>
      <Parameter name="minRowsPerProcessor"                 type="int" value="100"/>
      <Parameter name="nonzeroImbalance" type="double" value="1.1"/>
      <Parameter name="startLevel"                          type="int" value="1"/>
      <Parameter name="remapPartitions"                     type="bool" value="true"/>
    </ParameterList>

It's also important to set

   <Parameter name="maxCoarseSize"                         type="int" value="1"/>