Ifpack2: RILUK: Can now accept a RowMatrix that is not a CrsMatrix.
It used to be that RILUK would only take the input matrix A if it were
a Tpetra::CrsMatrix, even though the interface claims to accept any
Tpetra::RowMatrix.  This causes trouble if using RILUK as a subdomain
solver for Ifpack2::AdditiveSchwarz, since the latter often gives the
subdomain matrix to the subdomain solver as an Ifpack2::LocalFilter.
LocalFilter only implements RowMatrix, not CrsMatrix.  This commit
fixes the issue; RILUK now accepts any RowMatrix.

My fix is not the most optimal.  It simply copies the input matrix if
it is not a CrsMatrix.  Furthermore, my fix will not work correctly if
the matrix changes values between the initialize() and compute()
calls, since it copies the entire matrix in initialize().  Instead,
I'll need to copy just the graph in initialize(), then the values in
compute().  The latter can create the matrix with a const graph (from
initialize()), which should be faster.

I also fixed RILUK so that initialize() and compute() have correct
semantics (other than the caveat mentioned in the previous paragraph).
In particular, initialize() will now _always_ recompute
structure-related things, and compute() will now _always_ recompute
values-related things.  Before, these methods would only do anything
on their first call.  This was bad because users need to be able to
notify Ifpack2 that they changed the structure or values of the
matrix.

Build/Test Cases Summary
Enabled Packages: Ifpack2
Enabled all Forward Packages
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=23,notpassed=0 (3.77 min)
3) SERIAL_RELEASE_SS => passed: passed=18,notpassed=0 (6.18 min)