Adding support for inexact resid evaluations to ME::OutArgs::f (Trilinos #6015)
This should provide near perfect backward compatibility.  I have some unit
tests to demonstrate backward compatibility requirements and behavior when
slicing an RCP out of an Evaluation object.  A detailed design discussion is
contained in Trilinos #6015.

I will next change the response functions 'g' to use an Evaluation and update
all of the various decoarators and utility functions that depend on 'f' and
'g'.