Ifpack2: ReorderFilter: Fixed syntax error in my last commit.
The syntax error only manifests if both Xpetra and Zoltan2 are
enabled.  The latter means it might not show up in the nightly tests.

I also added a unit test which exercises AdditiveSchwarz with an RILUK
subdomain solver.  This test case was not exercised before.
Furthermore, if Xpetra and Zoltan2 are enabled, the test enables
reordering.  It's not a good correctness test, but it's a good test to
ensure that AdditiveSchwarz and RILUK play nicely together (i.e.,
neither throws exceptions, given correct parameters and input).

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2, Xpetra, Zoltan2
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=193,notpassed=0 (7.33 min)
3) SERIAL_RELEASE_SS => passed: passed=171,notpassed=0 (12.24 min)