Panzer: Adding ML coordinate support to the blocked assembly
This only worked for the monolithic assembly, this new parameter list
callback class provides an interface for it to work. Note in the
Teko input file, instead of putting "none" under the "Requested Parameters"
"x-coordinate" (etc...) you end up putting the name of the field that
the coordinates should be drawn from.