MueLu: slightly reordered ops in RepartitionFactory
We noticed that RepartitionFactory on level 1 takes a significan amount
of time even when the simplest "start level" check fails.

There were two relevant lines there:

1. Get<Matrix>("A", ...)
2. comm->duplicate()

The first one actually goes through the whole R,A,P, ... construction
process. Is not included in time, but what matters is that the process
may be imbalanced, so different processors exit the call at different
times.
The second one is the reason why we should never assume anything about
underlying libraries. Digging into internals of Teuchos MpiComm, it
turned out that any new MpiComm construction sets up some tag id, which
is then *synced* to all processors through *MPI_Bcast*.

So it's not RepartitioningFactory that takes that long, it's the sync of
all processors there.

I've rearranged some operations so that duplicate call happens only
after successful passing of the first check, when we would need to sync
anyway.