Amesos2:  Fix NVCC invalid cast errors.
NVCC doesn't appear to like the static casts in the Tpetra and Epetra
matrix adapters, even though the casts appear valid to me.  So for NVCC
I replaced them with dynamic casts.  It isn't clear to me if this is
really the correct thing to do as the classes don't seem to be run-time
polymorphic.