MueLu: more amalgamation changes
GetGlobalDOFs was returning a std::vector.  This was really inefficient,
as its contents were then put into another std::vector.  I converted the
the method to return a single global ordinal, and moved the loop out
into the calling methods of AmalgamationInfo.  Also renamed
GetGlobalDOFs to GetGlobalDOF.

I did some vtune comparisons of calling GetGlobalDOF versus putting the
conversion directly in the calling methods.  It seems the compiler
inlines correctly, as the performance difference is negligible.