Printing required deps correctly (Trilinos #5957), added few more tests/examples
While running some of the example cases to document the TriBITS dependency
behavior, I noticed that TriBITS was printing "optional" for required
dependenices when enabling upstream packages.  This was noted in Trilinos
temp variable and all of the sudden, the list of enabled packages changed for
a few tests.  These test cases were not correctly enabling some optional
upstream packages when they should have been.  My guess is that the temp
variable hack I was using was turnning off the enable of some upstream
packages.  Macros are evil!

The other tests that had to be updated where those that did not print the
right thing when enabling upstream required packages.

Note that I was able to remove an entire package loop in this change.