stratimikos: adding try/catch block around belos adapter solve()
Some belos exceptions were not being caught in the stratimikos adapters - in particular the call to solve() was not wrapped in a try/catch block.  This resulted in failing the entire simulation instead of allowing for a time stepper to check the linear solve status for a failed solve so that it can cut the step and try again.  Belos exceptions are now caught and the solve status reported through the thyra linearOpWithSolve interface reports a failed solve.

Build/Test Cases Summary
Enabled Packages: Stratimikos
Disabled Packages: SEACAS,Sundance,PyTrilinos,Didasko,Stokhos,TrilinosCouplings,MueLu,STK
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=296,notpassed=0 (31.27 min)
1) SERIAL_RELEASE => passed: passed=291,notpassed=0 (12.36 min)
2) MPI_DEBUG_SS => passed: passed=375,notpassed=0 (20.60 min)