Find project-checkin-test-config.py under --src-dir first, document TRIBITS_CHECKIN_TEST_DEBUG_DUMP
The big change here was was to change the checkin-test.py script to first look
for project-checkin-test-config.py under the project's source dir instead of
under tribitsDir/../.. which assumes that tribits is snapshotted into the
project sources.  I don't have a unit test for this yet but I can add one.
This feature only affects when --src-dir is explicitly passed in.

I also added a note to --help about setting TRIBITS_CHECKIN_TEST_DEBUG_DUMP=ON
for additional debugging output.