Belos: MINRES: "Num Blocks" parameter now allowed for compatibility.
Many Belos solvers based on GMRES have a "Num Blocks" parameter.  This
gives the restart length.  (The name refers to the amount of vector
storage, as a multiple of the solver's block size, but it really means
"restart length.")  Thus, a reasonable default ParameterList for Belos
may have the "Num Blocks" parameter in it.  However, Belos' MINRES
implementation currently throws an exception if given this parameter.
This makes sense, because its MINRES does not restart.  However, it
causes trouble for users who might want to reuse a ParameterList for
different Belos solvers.  This commit changes Belos' MINRES
implementation so that it accepts, but ignores, the "Num Blocks"
parameter.  I made this change in support of my recent added features
to the TrilinosCouplings scaling examples, so that it's possible to
use MINRES if changes to the material tensor make the matrix
indefinite.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Intrepid, Pamgen, Epetra, AztecOO, ML, TrilinosCouplings
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_RELEASE => passed: passed=359,notpassed=0 (7.69 min)